backend/requirements.txt
Django==5.1.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
sendgrid==6.11.0
celery==5.3.4
redis==5.0.1
psycopg[binary]==3.2.3
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0

# HTTP client for external API calls (Kickbox)
requests==2.32.3



